old/requirements.txt
numpy==1.26.4
sounddevice==0.4.6
webrtcvad==2.0.10                        # backup VAD only
onnxruntime==1.18.1
vosk==0.3.45
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
uvloop==0.19.0; platform_system != "Windows"
orjson==3.10.7
Pillow==10.3.0
# waveshare panel driver (official library with lgpio support)
epd-library
spidev
lgpio
# tokenizer + hf tokenizer runtime (CPU only; no torch required)
transformers==4.40.2
tokenizers>=0.19,<0.20